+++ src/components/content/timeline/patient-timeline.tsx
@@ -2,6 +2,7 @@ import cx from "classnames";
 import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
 import { patientTimelineColumns } from "./patient-timeline-columns";
 import { CodingList } from "@/components/core/coding-list";
+import { useCTW } from "@/components/core/providers/ctw-provider";
 import { Table } from "@/components/core/table/table";
 import { usePatientEncounters } from "@/fhir/encounters";
 import { EncounterModel } from "@/fhir/models/encounter";
@@ -12,11 +13,9 @@ export type PatientTimelineProps = {
   includeViewFhirResource?: boolean;
 };
 
-export function PatientTimeline({
-  className,
-  includeViewFhirResource = false,
-}: PatientTimelineProps) {
+export function PatientTimeline({ className }: PatientTimelineProps) {
   const patientEncounterQuery = usePatientEncounters();
+  const { featureFlags } = useCTW();
   const openDetails = useResourceDetailsDrawer({
     header: (m) => `${m.periodStart} - ${m.periodEnd}`,
     subHeader: (m) => m.typeDisplay,
@@ -35,7 +34,7 @@ export function PatientTimeline({
       <Table
         isLoading={patientEncounterQuery.isLoading}
         records={encounters}
-        columns={patientTimelineColumns(includeViewFhirResource)}
+        columns={patientTimelineColumns(featureFlags?.enableViewFhirButton)}
         handleRowClick={openDetails}
       />
     </div>